Fix backlog warnings in fusion
Previously, fusion_thread would send synced messages to the
fusion_server thread via a thread-safe Queue. fusion_server would poll
queue every 0.2 seconds to see if there is a new synced message. This
has now been changed to block instead until there is a message.

The reason this seems to help is get() needs to acquire locks, which
seems to be more expensive in Python 3 than in Python 2. This may be
the reason the backlog in the queue was growing too large. Waiting for
shorter time (0.02) didn't help with backlog warnings, which suggests
that the bottleneck is the time spent acquiring locks than actual
waiting.