Rewrite GrabPlaceModule in terms of collider bounds
Previous version of GrabPlaceModule had manually chosen thresholds for
creating intermediate points for hand motions. The current
implementation replaces these thresholds largely by relying on the
collider bounds instead. Some thresholds are still kept to configure how
high to lift the block, how close to reach the block before descending,
etc. A limitation of this approach may be that collider bounds should be
symmetrical to actual object mesh, so that it's center coincides with
the geometrical position of the object.