Rework DianaMotorControl to allow controlling wave animation
This will allow DianaMotorControl to control more animations as added.
The main change that allows this is that DianaMotorControl now responds
to me:intent:motion:rightArm for actually starting the animations. This
way me:intent:handPosR only serves to change animator target position
when the agent is not reaching, and to move the hand in
linear motion when the agent is reaching or has reached the target.