Make grasp action interruptible by ungrasp
Earlier, the animator controller had different states for Grab and
Ungrab. Grab is a blend animation, and Ungrab is its copy with speed set
to -1, hence the "ungrabbing" animation. This change removes the Ungrab
state, replaces the boolean "grab" animator parameter with float "grab".
This parameter now serves as the speed modifier for the Grab state as
well as a boolean by way of being either 1.0 or -1.0. It also makes the
Idle -> Grab transition interruptible by Grab -> Idle transition. To
ensure that DianaMotorControl and IdleBehaviour work as before, any
previous interaction with "grab" parameter has been suitably adjusted.